Refatoracao e criacao de novos testes no ffpojo-samples
