+++ ffpojo-samples/src/main/java/com/github/ffpojo/samples/SimpleDelimitedRecordParsingWriteExample.java
@@ -6,45 +6,41 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 import com.github.ffpojo.FFPojoHelper;
 import com.github.ffpojo.exception.FFPojoException;
-import com.github.ffpojo.exception.FieldDecoratorException;
-import com.github.ffpojo.metadata.DefaultFieldDecorator;
-import com.github.ffpojo.metadata.FieldDecorator;
-import com.github.ffpojo.metadata.positional.annotation.PositionalField;
-import com.github.ffpojo.metadata.positional.annotation.PositionalRecord;
+import com.github.ffpojo.metadata.delimited.annotation.DelimitedField;
+import com.github.ffpojo.metadata.delimited.annotation.DelimitedRecord;
 
-public class PositionalRecordParsingWithDecoratorExample {
+public class SimpleDelimitedRecordParsingWriteExample {
 
-	private static final String INPUT_TXT_RESOURCE_CLASSPATH = "PositionalRecordParsingWithDecoratorExampleInput.txt";
+	private static final String INPUT_TXT_RESOURCE_CLASSPATH = "SimpleDelimitedRecordParsingExampleInput.txt";
 	
 	//change here (make sure you have permission to write in the specified path):
-	//private static final String OUTPUT_TXT_OS_PATH = "C:/Users/gibaholms/Desktop/PositionalRecordParsingWithDecoratorExampleOutput.txt";
-	private static final String OUTPUT_TXT_OS_PATH = "C:/Users/gilberto.holms/Documents/temp/PositionalRecordParsingWithDecoratorExampleOutput.txt";
+	private static final String OUTPUT_TXT_OS_PATH = System.getProperty("java.io.tmpdir") +  INPUT_TXT_RESOURCE_CLASSPATH;
 	
-	@PositionalRecord
+	@DelimitedRecord(delimiter = "|")
 	public static class Customer {
 
 		private Long id;
 		private String name;
 		private String email;
-		private Date birthDate;
 		
-		@PositionalField(initialPosition = 1, finalPosition = 5, decorator = LongDecorator.class)
+		@DelimitedField(positionIndex = 1)
 		public Long getId() {
 			return id;
 		}
 		public void setId(Long id) {
 			this.id = id;
 		}
+		// must use a String setter or a FieldDecorator
+		public void setId(String id) {
+			this.id = Long.valueOf(id);
+		}
 		
-		@PositionalField(initialPosition = 6, finalPosition = 25)
+		@DelimitedField(positionIndex = 2)
 		public String getName() {
 			return name;
 		}
@@ -52,49 +48,17 @@ public class PositionalRecordParsingWithDecoratorExample {
 			this.name = name;
 		}
 		
-		@PositionalField(initialPosition = 26, finalPosition = 55)
+		@DelimitedField(positionIndex = 3)
 		public String getEmail() {
 			return email;
 		}
 		public void setEmail(String email) {
 			this.email = email;
 		}
-		
-		@PositionalField(initialPosition = 56, finalPosition = 65, decorator = DateDecorator.class)
-		public Date getBirthDate() {
-			return birthDate;
-		}
-		public void setBirthDate(Date birthDate) {
-			this.birthDate = birthDate;
-		}
-	}
-	
-	public static class LongDecorator extends DefaultFieldDecorator {
-		@Override
-		public Object fromString(String str) {
-			return Long.valueOf(str);
-		}
-	}
-	
-	public static class DateDecorator implements FieldDecorator<Date> {
-		private static final SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy");
-		
-		public Date fromString(String str) throws FieldDecoratorException {
-			try {
-				return sdf.parse(str);
-			} catch (ParseException e) {
-				throw new FieldDecoratorException("Error while parsing date field from string: " + str, e);
-			}
-		}
-
-		public String toString(Date obj) {
-			Date date = (Date)obj;
-			return sdf.format(date);
-		}
 	}
 	
 	public static void main(String[] args) {
-		PositionalRecordParsingWithDecoratorExample example = new PositionalRecordParsingWithDecoratorExample();
+		SimpleDelimitedRecordParsingWriteExample example = new SimpleDelimitedRecordParsingWriteExample();
 		try {
 			System.out.println("Making POJO from TXT...");
 			example.readCustomersFromText();
@@ -116,7 +80,7 @@ public class PositionalRecordParsingWithDecoratorExample {
 		String line;
 		while ( (line = textFileReader.readLine()) != null) {
 			Customer cust = ffpojo.createFromText(Customer.class, line);
-			System.out.printf("[%d][%s][%s][%s]\n", cust.getId(), cust.getName(), cust.getEmail(), cust.getBirthDate());
+			System.out.printf("[%d][%s][%s]\n", cust.getId(), cust.getName(), cust.getEmail());
 		}
 		textFileReader.close();
 	}
@@ -144,7 +108,6 @@ public class PositionalRecordParsingWithDecoratorExample {
 			cust.setId(98456L); 
 			cust.setName("Axel Rose"); 
 			cust.setEmail("axl@thehost.com");
-			cust.setBirthDate(new Date());
 			customers.add(cust);
 		}
 		{
@@ -152,7 +115,6 @@ public class PositionalRecordParsingWithDecoratorExample {
 			cust.setId(65478L); 
 			cust.setName("Bono Vox"); 
 			cust.setEmail("bono@thehost.com");
-			cust.setBirthDate(new Date());
 			customers.add(cust);
 		}
 		{
@@ -160,7 +122,6 @@ public class PositionalRecordParsingWithDecoratorExample {
 			cust.setId(78425L); 
 			cust.setName("Bob Marley"); 
 			cust.setEmail("marley@thehost.com");
-			cust.setBirthDate(new Date());
 			customers.add(cust);
 		}
 		return customers;
