+++ ffpojo-samples/src/main/java/com/github/ffpojo/samples/pojo/Header.java
@@ -1,0 +1,34 @@
+package com.github.ffpojo.samples.pojo;
+
+import java.util.Date;
+
+import com.github.ffpojo.decorator.DateDecorator;
+import com.github.ffpojo.metadata.positional.annotation.PositionalField;
+import com.github.ffpojo.metadata.positional.annotation.PositionalRecord;
+
+@PositionalRecord
+public class Header {
+
+	private Long controlNumber;
+	private Date processDate;
+	
+	@PositionalField(initialPosition = 1, finalPosition = 10)
+	public Long getControlNumber() {
+		return controlNumber;
+	}
+	public void setControlNumber(Long controlNumber) {
+		this.controlNumber = controlNumber;
+	}
+	// must use a String setter or a FieldDecorator
+	public void setControlNumber(String controlNumber) {
+		this.controlNumber = Long.valueOf(controlNumber);
+	}
+	
+	@PositionalField(initialPosition = 11, finalPosition = 20, decorator = DateDecorator.class)
+	public Date getProcessDate() {
+		return processDate;
+	}
+	public void setProcessDate(Date processDate) {
+		this.processDate = processDate;
+	}
+}
