--- ffpojo/src/main/java/com/github/ffpojo/AddClassBuilder.java
@@ -1,11 +1,0 @@
-package com.github.ffpojo;
-
-import java.util.List;
-
-public interface AddClassBuilder {
-
-	public FFReaderBuilder addRecordClass(Class<?> clazz);
-	
-	public FFReaderBuilder addRecordClasses(List<Class<?>> clazz);
-	
-}
