--- ffpojo/src/main/java/com/github/ffpojo/FFReaderBuilder.java
@@ -1,9 +1,0 @@
-package com.github.ffpojo;
-
-import java.util.List;
-
-public interface FFReaderBuilder {
-
-	public List<?> read ();
-	
-}
