+++ ffpojo/src/main/java/com/github/ffpojo/dsl/AddClassBuilder.java
@@ -1,0 +1,12 @@
+package com.github.ffpojo.dsl;
+
+import java.util.List;
+
+@SuppressWarnings("all")
+public interface AddClassBuilder {
+
+	public FFReaderBuilder withRecordClass(Class<?> clazz);
+	
+	public FFReaderBuilder withRecordClasses(List<Class<?>> clazz);
+	
+}
