+++ ffpojo/src/main/java/com/github/ffpojo/dsl/InputStreamReaderBuilder.java
@@ -1,4 +1,4 @@
-package com.github.ffpojo;
+package com.github.ffpojo.dsl;
 
 public interface InputStreamReaderBuilder {
